Add rest-server and soap src files to cmake to include in gtest. Let's get testing! - Merge rest-server/CMakeLists.txt into test/CMakeLists.txt - Add SOAP src files to test/CMakeLists.txt for testing - Refactor XMLParser.h to .hpp - Remove main() from XMLParser.cpp
