+++ SOAP/XML-parsing/Functionality Tests/UsersXMLParseTest.cpp
@@ -14,7 +14,7 @@
 #include <string>
 #include <fstream>
 #include <vector>
-#include "../XMLParser.h"
+#include "../XMLParser.hpp"
 #include <boost/foreach.hpp>
 #include <boost/property_tree/xml_parser.hpp>
 #include <boost/property_tree/ptree.hpp>
